+++ docs/source/changelog.md
@@ -5,6 +5,14 @@
 :width: 200px
 :align: center
 ```
+## 3.7.0
+1. Added WebSocket support
+   - Various polling elements from the frontend has been replaced with WebSocket connections for real-time updates
+2. BREAKING CHANGE: Removed `instances.py` module
+    - Plugins and other modules should now access the `AppContext` object via dependency injection in FastAPI
+3. Fixed Custom Downloads installs
+4. Various dependency updates
+
 ## 3.6.3
 
 ### New Features: